Various improvements to Docker daemon start/stop logic
When starting Supervisor often the terminal messed up. From
investigation it seemed that dockerd messes with the terminal somehow,
carriage returns (\r) seem not to function properly. Resetting the
terminal fixes this.

While add it, also use the Docker daemon PID from the pid file. This
is a different PID than before, since this is the actual daemon PID and
not the PID of the sudo process that starts the daemon. This should
not really make a difference, since killing the sudo process seemed
to have passed the signal along to the daemon, but it is cleaner to
use the actual daemon PID.