Use systemd-journal as logging backend for Docker daemon (#118)
* Use systemd-journal as logging backend for Docker daemon

Use systemd-journal as logging backend for Docker daemon so that the
advanced logging works in development environment too. This currently
needs a work around to have systemd-journal-gatewayd available as a
unix domain socket since we don't have systemd units which can provide
the socket activation for the systemd-journal-gatewayd daemon.

* Use systemd-journal-gatewayd URL instead of socat

Drop socat in favor of the new environment variable to pass a URL to
connect to the systemd-journal-gatewayd.

* Add daemon.json file to configure docker to use systemd-journald logging